Rename plans/analyser_scans to plans/analyser_scan plust test
