+++ src/sm_bluesky/electron_analyser/plans/__init__.py
@@ -1,3 +1,3 @@
-from .analyser_scans import analysercount, analyserscan, grid_analyserscan
+from .analyser_scan import analysercount, analyserscan, grid_analyserscan
 
 __all__ = ["analysercount", "analyserscan", "grid_analyserscan"]
